Unify LLMCongif and EmbeddingConfig
