--- tt-media-server/cpp_server/include/config/embedding_config.hpp
@@ -1,14 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
-
-#pragma once
-
-namespace tt::config {
-
-/**
- * Configuration for embedding service.
- * Currently a placeholder - will be expanded as embedding features are added.
- */
-struct EmbeddingConfig {};
-
-}  // namespace tt::config
