+++ tt-media-server/cpp_server/include/config/settings.hpp
@@ -4,8 +4,7 @@
 #pragma once
 
 #include "config/types.hpp"
-#include "config/llm_config.hpp"
-#include "config/embedding_config.hpp"
+#include "config/runner_config.hpp"
 #include <cstddef>
 #include <string>
