+++ tt-media-server/cpp_server/include/ipc/boost_ipc_task_queue.hpp
@@ -9,7 +9,7 @@
 #include <boost/interprocess/ipc/message_queue.hpp>
 
 #include "runners/llm_runner/task_queue.hpp"
-#include "config/llm_config.hpp"
+#include "config/runner_config.hpp"
 
 namespace tt::ipc {
