+++ tt-media-server/cpp_server/tests/ipc_scheduler_smoke_test.cpp
@@ -12,7 +12,7 @@
 #include <boost/uuid/uuid.hpp>
 #include <boost/uuid/uuid_generators.hpp>
 #include <boost/uuid/uuid_io.hpp>
-#include "config/llm_config.hpp"
+#include "config/runner_config.hpp"
 #include "ipc/boost_ipc_task_queue.hpp"
 #include "runners/llm_runner/scheduler.hpp"
 #include "runners/llm_runner/prefill_first_scheduler.hpp"
